backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.136.0",
    "pgvector>=0.4.2",
    "pwdlib[argon2]>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
]
